backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
sqlmodel>=0.0.11
alembic>=1.12.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
redis>=5.0.1
celery>=5.3.4
structlog>=23.2.0
prometheus-fastapi-instrumentator>=6.1.0
sentry-sdk>=1.32.0
fastapi-limiter>=0.1.5
python-dotenv>=1.0.0
pydantic[email]
email-validator
